http_*: put http api relevant code in a new file
